move: Rename cubeState to boxModel
